// src/app/features/auth/interfaces/IAuthResponse.ts
export interface User {
  _id: string;
  name: string;
  username: string;
  email: string;
  photo: string;
  cover: string;
}

export interface Data {
  token: string;
  tokenType: string;
  expiresIn: string;
  user: User;
}

export interface IAuthResponse {
  success: boolean;
  message: string;
  data: Data;
}